Improve behaviour opening a project in vim
Previously, the directory would open up with NERDTree in the main
window, but on opening up a file the tree would stay in place until
`:q`ing it, and hitting <F2> would open up a second NERDTree. By
changing the default behaviour (on the `v` alias) to open vim with no
arguments, and changing vim's behaviour to open NERDTree if no arguments
are specified, NERDTree opens up on the left-hand side, and I can
dismiss it easily once I've found the file I want.