Improve bash prompt, now with more colour
I've removed the hostname and username from my prompt -- if I'm on my
machine, I don't need to know which user I am -- and added some colour
to show which part is the directory. I've also added the current Git
branch.

One thing I was hoping to do is to change the colour of the '$' at the
end to reflect whether the last-run command succeeded or failed, but I
wasn't able to get this working with the various `function`s I call,
each of which appears to reset the exit code when it runs.

I renamed the `pairing-prompt` file to `prompt-pairing`, to get it to
load after the `prompt` file -- it's a slight hack, but it keeps the two
separate.