Remove old non-subtreed version of fugitive
When I added this, it was before I knew about the `git subtree` command
(and possibly before it even existed), so the commit isn't in a format
that `git subtree` recognises.

Given I haven't modified this plugin in any way, there's no real harm in
blatting it and starting afresh.