Remove `se` function to activate virtualenvs
These days, I almost always use either `virtualenvwrapper` (which gives
me the `workon` command to switch to a directory and activate its
environment), or `direnv` (which activates the environment as part of
`cd`-ing to the directory).

For those very few cases where I don't do this, I'm happy typing out
`source bin/activate`, or whatever the equivalent path is.