Add config for zsh
Having started at a company that has a bunch of extra integrations with
zsh, now seems as good a time as ever to give it a go. Seems promising
so far.

This is more or less a direct port of my existing bash config, so I'm
sure I'm only scratching the surface of what's possible with zsh, but
this lets me get up and running without having to go all-in on something
like oh-my-zsh and spend days elbow-deep in plugin libraries.