Add Black support
This plugin is from commit d33337b4d5 of https://github.com/psf/black;
it seemed easier to copy-paste it here than to faff around with another
subtree merge. It's not like I'm planning to change the plugin or
anything.

I'm leaving it as an explicit command for now, although I'm sure there's
a way to set it up so that it runs automatically for certain projects.