Switch `git com` to check out the `main` branch
I'm using this as the default main branch for whichever repos I can,
because it's easy to change and a good habit to get into. For those
repos that are harder to change (because some tooling still confers
special status on the `master` name), I can override the command on a
per-repo basis. I'm ok with having a little bit of extra friction and an
extra reminder in those cases.