Use the pipx virtualenv to load Black
Black's default Vim integration creates a virtualenv at `~/.vim/black`,
installs Black inside it, and uses that. Now that I'm using pipx to
manage command-line Python utilities, it seems useful to have there be
one canonical version of Black running, and have one tool responsible
for managing it. Either that or have it as an explicit project
dependency, to avoid changes between versions, but that doesn't seem to
be easy to do with the stock plugin, and I don't particularly fancy
writing my own.

This feels a little hacky, as my Vim config now depends on the details
of my pipx config, but in the case of Black, consistency outweighs
isolation.